Add OpenSearch
> An OpenSearch description document can be used to describe the web
interface of a search engine.

http://www.opensearch.org/Specifications/OpenSearch/1.1

The [idea][1] here was to get Chrome to automatically pick up the search
and add it to its list of search engines, so you could just type this in
the browser's location bar:

    go <word>

... and it would autocomplete the search right in the omnibox.

I can't get it to work in Chrome, though. Maybe it's fixable.

[1]: http://stackoverflow.com/questions/1317051/making-usage-of-google-chromes-omnibox-tab-feature-for-on-personal-website