Make app deployable with Tornado
Tornado is an open source version of the scalable, non-blocking web
server and tools that power FriendFeed. Because it is non-blocking and
uses epoll, it can handle thousands of simultaneous standing
connections, which means it is ideal for real-time web services.

Exactly what we need for this goto app.

Also, add a signal handler for SIGTERM to allow for proper killing with
service scripts.